+++ frontend/src/App.jsx
@@ -27,8 +27,8 @@ import { MemberLogout } from "./feature/Member/MemberLogout.jsx";
 import { ProductOrderComplete } from "./feature/Product/ProductOrderComplete.jsx";
 import { CartProvider } from "./feature/Product/CartContext.jsx";
 import Main from "./feature/common/Main.jsx";
-import ImageSlide from "./feature/common/ImageSlide.jsx";
-import Home from "./feature/common/Home.jsx";
+import ImageSlide from "./feature/common/CoverImageSlide/ImageSlide.jsx";
+import Home from "./feature/common/Home/Home.jsx";
 
 function App() {
   // const username = prompt("닉네임을 입력해 주세요");
