+++ frontend/src/feature/common/Home/CategoryMenu.css
@@ -1,0 +1,34 @@
+.category-menu-wrapper {
+    padding: 2rem 1rem;
+    background-color: white;
+}
+
+.category-title {
+    font-size: 1.5rem;
+    font-weight: 600;
+    margin-bottom: 1rem;
+}
+
+.category-list-scroll {
+    display: flex;
+    overflow-x: auto;
+    gap: 1.5rem;
+    padding-bottom: 1rem;
+}
+
+.category-item {
+    flex: 0 0 auto;
+    text-align: center;
+    cursor: pointer;
+}
+
+.category-icon {
+    font-size: 2.5rem;
+    margin-bottom: 0.5rem;
+}
+
+.category-label {
+    font-size: 0.9rem;
+    color: #333;
+    white-space: nowrap;
+}
